+++ docs/source/changelog.rst
@@ -11,6 +11,7 @@ Version (dev)
 * Add :mod:`srctools.sndscape`, for parsing soundscape scripts.
 * Changed `PackList.load_choreo_scenes() <srctools.packlist.PackList.load_choreo_scenes>` to warn if ``scenes.image`` is not found.
 * Add support for ``orientedwidthheight()``, ``orientedwidthheighthalf()`` and ``clusteredlight()`` Strata FGD helpers.
+* Added ``noinherit()`` FGD helper, for indicating values to avoid inheriting from base classes.
 
 -------------
 Version 2.6.2